Added new type RowItemActions
A new type 'actions' has been added to the group <rows> where you can indicate buttons for actions and modal windows.
